requirements.txt
# YCCE-AI Dependencies
langchain>=0.3.0
langchain-text-splitters>=0.3.0
supabase>=2.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
PyMuPDF>=1.24.0
python-dotenv>=1.0.0
lxml>=5.0.0
